+++ R/checkIntOverflow.R
@@ -9,7 +9,7 @@
 #' This restriction is relevant when exporting a \code{GADSdat} to \code{.dta}
 #' and only when any value exceeding the limit is also labeled (or tagged as missing).
 #' This is because Stata only accepts labeled \emph{integer} (not labeled \emph{floating-point};
-#' c.f. \code{\link[checkLabeledFractionals()]{checkLabeledFractionals}} in this package)
+#' c.f. \code{\link{checkLabeledFractionals}()} in this package)
 #' values. \code{haven}'s \code{\link[haven:write_dta]{write_dta}} function will therefore
 #' try to coerce any labeled values \code{as.integer()}. Unlabeled values, however, will
 #' stay generic \code{numeric} values that have a higher limit.
@@ -21,6 +21,8 @@
 #' and whether they actually occur in the data (\code{empty}).
 #' The \code{rownum}s of the affected rows in \code{GADSdat$labels} are also
 #' provided in a separate column as a fail safe.
+#'
+#'@family dataset compliance checks
 #'
 #'@examples
 #' # Introduce a fractional value into meta data