+++ NEWS.md
@@ -1,12 +1,17 @@
 # eatGADS 1.2.0.9000
+## new features
+* `checkIntOverflow()` and `checkLabeledFractionals()` check if a `GADSdat` complies with Stata's requirements for data sets (#110)
+* `checkVarNames()` can now also check the lengths of variable names against program-specific restrictions (#110)
+* A list of limits to dataset components under Stata and SPSS is available in `program_limits` and via its access function `getProgramLimit()`.
+
 ## bug fixes
 * `autoRecode()` no longer fails while applying a template that does not yet cover a value which occurs in the data more than once (#131)
 * `changeMissings()` no longer throws an existing value error in edge cases with multiple existing value labels
 * `recodeGADS()` now correctly performs potentially sequential recodes for unlabeled values (#134)
+* arguments in `checkVarNames()` now turn on/off different checks correctly (#138)
 
 ## documentation
 * The behavior of `autoRecode()` is further clarified, emphasizing the dropping of value labels and pointing to `multiChar2fac()` (#136).
-
 
 # eatGADS 1.2.0
 ## new features